upstream: b=master,r=1b92cff5c5610e73df54724e81dbf0b519830fba,t=2015-09-24-2046-24584

+++ components/nexus-httpclient/src/main/java/org/sonatype/nexus/httpclient/config/SecondsDeserializer.java
@@ -28,15 +28,12 @@ import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 public class SecondsDeserializer
     extends StdDeserializer<Time>
 {
-
   public SecondsDeserializer() {
     super(Time.class);
   }
 
   @Override
-  public Time deserialize(final JsonParser parser, final DeserializationContext context)
-      throws IOException
-  {
-    return Time.seconds(parser.readValueAs(Integer.class));
+  public Time deserialize(final JsonParser parser, final DeserializationContext context) throws IOException {
+    return Time.seconds(parser.readValueAs(Long.class));
   }
 }
